requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-environ==0.11.2
django-role-permissions==3.2.0
greenlet==3.0.3
idna==3.7
psycopg2==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
